service, diagnostics and surveys are only doctor case property
