+++ app/AccidentAbstract.php
@@ -45,37 +45,6 @@ abstract class AccidentAbstract extends Model
     }
 
     /**
-     * Selected by doctor diagnostics
-     */
-    public function diagnostics(): MorphToMany
-    {
-        return $this->morphToMany(Diagnostic::class, 'diagnosticable');
-    }
-
-    /**
-     * Each DoctorAccident is able to has own services, created by a doctor
-     * but by default it could be defined by the director
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphToMany
-     */
-    public function services(): MorphToMany
-    {
-        return $this->morphToMany(DoctorService::class, 'doctor_serviceable');
-    }
-
-    /**
-     * As same as serviceable()
-     * each doctorAccident is able to has his own survey
-     * but by default it could be defined by the director
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphToMany
-     */
-    public function surveys(): MorphToMany
-    {
-        return $this->morphToMany(DoctorSurvey::class, 'doctor_surveable');
-    }
-
-    /**
      * Photos of the documents from the patient
      *
      * @return \Illuminate\Database\Eloquent\Relations\MorphToMany
