+++ app/Http/Controllers/Api/V1/Director/CasesController.php
@@ -225,9 +225,9 @@ class CasesController extends ApiController
             if ($accident->isDoctorCaseable()) {
                 $caseableAccidentData = $request->json('doctorAccident', []);
                 // attach services, surveys and diagnostics
-                $this->updateDoctorMorph($accident->caseable, $request, 'services');
-                $this->updateDoctorMorph($accident->caseable, $request, 'surveys');
-                $this->updateDoctorMorph($accident->caseable, $request, 'diagnostics');
+                $this->updateDoctorMorph($accident->getAttribute('caseable'), $request, 'services');
+                $this->updateDoctorMorph($accident->getAttribute('caseable'), $request, 'surveys');
+                $this->updateDoctorMorph($accident->getAttribute('caseable'), $request, 'diagnostics');
             } else {
                 $caseableAccidentData = $request->json('hospitalAccident', []);
             }
@@ -302,18 +302,6 @@ class CasesController extends ApiController
     private function updateDoctorMorph(DoctorAccident $doctorAccident, Request $request, string $morphName): void
     {
         $morphs = $request->json($morphName, []);
-        /*
-         * I need to go through the new parameters, not the already stored
-         * (definitely for the services)
-         *
-         * not clear why do I need it at all
-         * if ($doctorAccident->$morphName()) {
-            foreach ($doctorAccident->$morphName() as $morph) {
-                if ($morph && in_array($morph->id, $morphData, false) ) {
-                    $morphs[] = $morph->id;
-                }
-            }
-        }*/
         $doctorAccident->$morphName()->detach();
         $doctorAccident->$morphName()->attach($morphs);
     }
